+++ kalite/updates/management/commands/languagepackdownload.py
@@ -26,7 +26,7 @@ from fle_utils.general import ensure_dir
 from fle_utils.internet import callback_percent_proxy, download_file
 from kalite import caching
 from kalite.i18n import LOCALE_ROOT, DUBBED_VIDEOS_MAPPING_FILEPATH
-from kalite.i18n import get_localized_exercise_dirpath, get_srt_path, get_po_filepath
+from kalite.i18n import get_localized_exercise_dirpath, get_srt_path, get_po_filepath, get_language_pack_url
 from kalite.i18n import lcode_to_django_dir, lcode_to_ietf, update_jsi18n_file
 from kalite.version import SHORTVERSION
 
@@ -213,12 +213,3 @@ def move_srts(lang_code):
     else:
         logging.info("Removing empty source directory (%s)." % src_dir)
         shutil.rmtree(src_dir)
-
-
-def get_language_pack_url(lang_code, version=SHORTVERSION):
-    """As published"""
-    return "http://%(host)s/media/language_packs/%(version)s/%(lang_code)s.zip" % {
-        "host": settings.CENTRAL_SERVER_HOST,
-        "lang_code": lang_code,
-        "version": version,
-    }
